gen-ai/code-execution-with-mcp-strands-sdk/requirements.txt
strands-agents>=1.0.0
strands-agents-tools>=0.2.0
langfuse>=3.0.0
python-dotenv>=1.0.0